feat: unify storybook packages
