+++ package.json
@@ -62,12 +62,12 @@
     "schemas:generate": "bash scripts/generate-ts-schemas/generate-schemas.sh",
     "schemas:variables": "bash scripts/generate-ts-schemas/setup-openapi-ts.sh",
     "start": "pnpm run storybook",
-    "storybook": "concurrently --kill-others-on-fail 'pnpm run types:watch' 'nx run @integration-components/tools-storybook-preact:storybook'",
-    "storybook:vue": "nx run @integration-components/tools-storybook-vue:storybook",
-    "storybook:build": "nx run @integration-components/tools-storybook-preact:storybook:build",
-    "storybook:build:preact": "nx run @integration-components/tools-storybook-preact:storybook:build",
-    "storybook:build:vue": "nx run @integration-components/tools-storybook-vue:storybook:build",
-    "storybook:static": "nx run @integration-components/tools-storybook-preact:storybook:static",
+    "storybook": "concurrently --kill-others-on-fail 'pnpm run types:watch' 'nx run @integration-components/tools-storybook:storybook:preact'",
+    "storybook:vue": "concurrently --kill-others-on-fail 'pnpm run types:watch' 'nx run @integration-components/tools-storybook:storybook:vue'",
+    "storybook:build": "pnpm run storybook:build:preact",
+    "storybook:build:preact": "nx run @integration-components/tools-storybook:storybook:build:preact",
+    "storybook:build:vue": "nx run @integration-components/tools-storybook:storybook:build:vue",
+    "storybook:static": "nx run @integration-components/tools-storybook:storybook:static",
     "test": "cross-env TEST_ENV=1 vitest --config vite.config.ts",
     "test:contract": "cross-env NODE_OPTIONS='--no-experimental-strip-types' playwright test --project contract",
     "test:coverage": "vitest run --config vite.config.ts --coverage",
