--- packages/tools/storybook-preact/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "extends": "../../../tsconfig.base.json",
-    "compilerOptions": {
-        "noEmit": true,
-        "types": ["node"]
-    },
-    "files": ["../../../src/declarations/scss.d.ts", "../../../src/declarations/vite-env.d.ts", "../../../src/declarations/svgr.d.ts"],
-    "include": ["src/**/*.ts", "src/**/*.tsx", ".storybook/**/*.ts", ".storybook/**/*.tsx", "../../../stories/**/*.ts", "../../../stories/**/*.tsx"]
-}
