--- packages/tools/storybook-vue/.storybook/manager.ts
@@ -1,7 +1,0 @@
-import { addons } from 'storybook/manager-api';
-
-addons.setConfig({
-    isFullscreen: false,
-    showPanel: true,
-    panelPosition: 'bottom',
-});
