+++ packages/tools/storybook/.eslintrc.cjs
@@ -1,0 +1,13 @@
+module.exports = {
+    extends: '../../../.eslintrc.cjs',
+    rules: {
+        'import-x/no-extraneous-dependencies': [
+            'error',
+            {
+                devDependencies: true,
+                peerDependencies: true,
+                packageDir: ['./', '../../../'],
+            },
+        ],
+    },
+};
