+++ packages/tools/storybook/.storybook-preact/manager.ts
@@ -1,0 +1,1 @@
+import '../shared/manager';
