+++ packages/tools/storybook/.storybook-preact/tsconfig.json
@@ -4,5 +4,5 @@
         "moduleResolution": "bundler",
         "allowImportingTsExtensions": true
     },
-    "include": ["./**/*.ts", "./**/*.tsx", "../src/**/*.ts", "../src/**/*.tsx"]
+    "include": ["./**/*.ts", "./**/*.tsx", "../src/**/*.ts", "../src/**/*.tsx", "../shared/**/*.ts"]
 }
