+++ packages/tools/storybook/.storybook-vue/manager.ts
@@ -1,0 +1,1 @@
+import '../shared/manager';
