+++ packages/tools/storybook/.storybook-vue/preview.ts
@@ -1,0 +1,8 @@
+import type { Preview } from '@storybook/vue3';
+import { sharedPreviewConfig } from '../shared/previewDefaults';
+
+const preview: Preview = {
+    ...sharedPreviewConfig,
+};
+
+export default preview;
