+++ packages/tools/storybook/.storybook-vue/tsconfig.json
@@ -4,5 +4,5 @@
         "moduleResolution": "bundler",
         "allowImportingTsExtensions": true
     },
-    "include": ["./**/*.ts", "../src/**/*.ts", "../src/**/*.vue"]
+    "include": ["./**/*.ts", "../src/**/*.ts", "../src/**/*.vue", "../shared/**/*.ts"]
 }
