+++ packages/tools/storybook/shared/dev.js
@@ -1,14 +1,21 @@
 import { spawnSync } from 'child_process';
+import { existsSync, readFileSync } from 'fs';
 import { createConnection } from 'net';
 import { resolve, dirname } from 'path';
 import { fileURLToPath } from 'url';
-import { existsSync, readFileSync } from 'fs';
 import dotenv from 'dotenv';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const rootDir = resolve(__dirname, '../../../..');
 const projectRoot = resolve(__dirname, '..');
+const configDir = process.argv[2];
 
+if (!configDir) {
+    console.error('\nUsage: node shared/dev.js <config-dir>  (e.g. .storybook-preact)\n');
+    process.exit(1);
+}
+
+const label = configDir.replace(/^\.storybook-/, '');
 const envPath = existsSync(resolve(rootDir, 'envs/.env')) ? resolve(rootDir, 'envs/.env') : resolve(rootDir, 'envs/env.default');
 const parsed = dotenv.parse(readFileSync(envPath, 'utf8'));
 const port = parsed.PLAYGROUND_PORT || '3030';
@@ -21,12 +28,13 @@ const tryPort = host =>
             .once('connect', () => (sock.end(), r(true)))
             .once('error', () => r(false));
     });
+
 if ((await tryPort('127.0.0.1')) || (await tryPort('::1'))) {
-    console.error(`\n[storybook-preact] Port ${port} is already in use. Is the sibling Storybook still running?\n`);
+    console.error(`\n[storybook-${label}] Port ${port} is already in use. Is the sibling Storybook still running?\n`);
     process.exit(1);
 }
 
-const result = spawnSync('npx', ['storybook', 'dev', '-p', port, '--no-open', '--ci'], {
+const result = spawnSync('npx', ['storybook', 'dev', '-p', port, '--config-dir', configDir, '--no-open', '--ci'], {
     cwd: projectRoot,
     stdio: 'inherit',
     shell: true,
