+++ packages/tools/storybook/static/mockServiceWorker.js
