+++ packages/tools/storybook/vite.config.ts
@@ -1,8 +1,8 @@
 import { defineConfig } from 'vite';
 import { getEnvironment } from '../../../envs/getEnvs';
 
-// Minimal Vite config for `vite preview`. Storybook dev/build pipelines use the
-// config inlined in `.storybook/main.ts`.
+// Minimal Vite config for `vite preview` in `storybook:static`. The Storybook
+// dev/build pipelines use the config inlined in each .storybook-*/main.ts.
 const { app } = getEnvironment('development');
 
 export default defineConfig({
